atiny/requirements.txt
# http full
# MyRequestsHTTP
requests
requests[socks]
urllib3
PySocks

# aiohttp
aiohttp
aiosocksy
aiohttp_socks

# other
dnspython
ujson
responses
aresponses
websockets